requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# Linting and formatting
flake8==6.1.0
black==23.11.0
isort==5.12.0

# Debugging
flask-debugtoolbar==0.13.1
ipdb==0.13.13

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0